Started working on an annotator feature to label datasets for ML training.
This commig is the initial commit for this feature. It includes a feature flag to enable/disable the feature since it's
not ready for production yet. The feature is not yet complete, but the basic structure is in place.

To enable the feature, set `TRAINING_ANNOTATOR_FEATURE_FLAG=True` in `smartscope.conf`.